% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiXCR.R
\docType{package}
\name{MiXCR-package}
\alias{MiXCR}
\alias{MiXCR-package}
\title{StarSalmon: Functions for writing commands and post processing the Star Salmon pipeline.}
\description{
Given a sample data matrix that indicate the sample name, sample folder name and input paths
this script can write out pipeline commands to run star salmon on the cluster and post process the
data into ucsc_isoform_counts.tsv and hgnc_entrezid_gene_counts.tsv.  Only ucsc pipeline outputs 
are handled at this time.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/DanteBortone/MiXCR/issues}
}

}
\author{
\strong{Maintainer}: Dante Bortone \email{dantebortone@gmail.com}

}
\keyword{internal}
