% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_mixcr.R
\name{post_process_mixcr}
\alias{post_process_mixcr}
\title{post_process_mixcr}
\usage{
post_process_mixcr(input_file_paths, output_path, sample_data_table,
  my_chains = my_chains, sample_folder_column = "Sample_Folder",
  sample_id_column = "Sample_ID", thread_num = 1)
}
\arguments{
\item{input_file_paths}{Character vector of paths to the pipeline output data.}

\item{output_path}{Path to the output file}

\item{sample_data_table}{Data table which should contatin the sample_id_column and sample_folder_column}

\item{sample_folder_column}{The name of the column that has sample folder names}

\item{sample_id_column}{The name of the column that has sample ids}

\item{thread_num}{Integer number of threads to run mclapply statements}
}
\value{
A path to the rds file.
}
\description{
Joins individual sample files into one tsv file.  Also converts ucsc names to hgnc_symbol|entrez_ids
and outputs the gene level counts.
}
\details{
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
post_process_mixcr
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
}
