% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePostMiXCR.R
\docType{package}
\name{PrePostMiXCR-package}
\alias{PrePostMiXCR}
\alias{PrePostMiXCR-package}
\title{StarSalmon: Functions for writing commands and post processing the MiXCR pipeline.}
\description{
Given a sample data matrix that indicate the sample name and file paths post_process_mixcr will 
run the diversity calculation metrics on those MiXCR output files.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://bitbucket.org/unc_lineberger/pre-post-mixcr/issues}
}

}
\author{
\strong{Maintainer}: Dante Bortone \email{dantebortone@gmail.com}

}
\keyword{internal}
